fix(pricing): unit_price FLOAT -> NUMERIC(12,4), verified before and after
The top data-integrity item DEPLOY-001 left open. BR-0005 made all *arithmetic*
exact, but exact arithmetic on an inexact input is still inexact: the price was
already approximate before the first multiplication, and it is the number every
calculation, settlement, payment and receipt descends from.

NUMERIC(12,4) matches milk_collection_transaction.unit_price — the house
standard since PRC-004 and the column this value is copied into. A test now
fails if the two ever disagree.

THE CAST WAS PROVEN SAFE BEFORE THE MIGRATION WAS WRITTEN, because the answer
decides whether it is safe at all. PostgreSQL converts float8 to numeric
through the float's SHORTEST round-trip text form, not its binary expansion, so
44.7291 becomes exactly 44.7291 and not 44.729100000000002. That property is
what makes this a storage fix rather than a data-correction exercise.

The same experiment showed what the cast does NOT preserve: 44.72915 -> 44.7292,
0.00005 -> 0.0001, 0.00001 -> 0.0000 (which also violates ck_matrix_row_price).
So the migration counts any row with more than four decimals FIRST and aborts
naming them. Silently changing a price is precisely the failure this platform
exists to prevent.

VERIFIED BY EXECUTION, NOT ASSERTION. Nine representative prices written as
float8, migrated, compared row by row — all numerically identical; the
up -> down -> up round trip preserves every value; and a deliberately-planted
five-decimal price aborts the migration and leaves the column untouched.

CONTRACT CHANGE. unit_price now serialises as a STRING in API responses, like
every other money field on this API (payment.amount, settlement.net_amount) —
the tests that asserted a JSON number were encoding the old inconsistency.
Requests may still send a number; Pydantic parses either into Decimal. The
admin portal and the Flutter client both needed updating: the Dart client did
`(json['unit_price'] as num).toDouble()`, which would have THROWN on a string.

from_value/to_value deliberately stay FLOAT. They are band BOUNDARIES compared
against a quality reading, not money; converting them changes which band a
boundary-valued reading selects, which is a behavioural change to BR-0004 and
needs its own analysis rather than a ride along with a storage fix.

New PostgreSQL-only suite test_pricing_precision_postgres.py (5 tests), wired
into postgres-proof.sh — SQLite has no numeric type with a declared scale and
would accept every assertion while storing a float.

1,094 tests collected, SQLite suite green, portal typecheck/build/eslint clean,
flutter analyze clean with 49 tests passing. Divergence #61 closed.

+++ services/platform-core/src/platform_core/modules/pricing/matrix.py
@@ -12,6 +12,7 @@ the reviewed pricing data is exactly what publishes.
 import itertools
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from pydantic import BaseModel, Field, field_validator, model_validator
 from sqlalchemy import func, or_, select
@@ -128,7 +129,11 @@ class UpdateMatrixCommand(BaseModel):
 class RowInput(BaseModel):
     from_value: float
     to_value: float
-    unit_price: float = Field(gt=0)
+    #: Decimal from the edge inward. Pydantic parses a JSON number into Decimal
+    #: via its string form, so `44.7291` arrives as Decimal("44.7291") rather
+    #: than as the float that is really 44.729100000000002 — the price is exact
+    #: before it is ever stored, not merely exact once arithmetic starts.
+    unit_price: Decimal = Field(gt=0)
     sequence: int | None = Field(default=None, ge=1)
     active: bool = True
 
@@ -159,7 +164,7 @@ class RowView(BaseModel):
     sequence: int
     from_value: float
     to_value: float
-    unit_price: float
+    unit_price: Decimal
     active: bool
 
     model_config = {"from_attributes": True}
